requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.2.0
langchain-mistralai
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.2.0